[KDL005] #110215 chinh sua lai code theo EA

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/find/dialog/employmentsystem/PegManagementDto.java
@@ -14,7 +14,7 @@ public class PegManagementDto {
 	private double usageDay;
 	
 	// field 使用時間数
-	private int usageHour;
+	private double usageHour;
 	
 	// field 開発日
 	private GeneralDate developmentDate;
